Refactor app_bsky::actor
